requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
pandas==2.2.1
python-dotenv==1.0.1
openai==1.12.0
lxml==5.1.0
webdriver-manager==4.0.1
